+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/australia/__init__.py
@@ -2,12 +2,14 @@
 
 from .au_abn_recognizer import AuAbnRecognizer
 from .au_acn_recognizer import AuAcnRecognizer
+from .au_bank_details_recognizer import AuBankDetailsRecognizer
 from .au_medicare_recognizer import AuMedicareRecognizer
 from .au_tfn_recognizer import AuTfnRecognizer
 
 __all__ = [
     "AuAbnRecognizer",
     "AuAcnRecognizer",
+    "AuBankDetailsRecognizer",
     "AuMedicareRecognizer",
     "AuTfnRecognizer",
 ]